services/data-ingestor/requirements.txt
paho-mqtt==2.1.0
pymysql==1.1.1
